terraform/src/requirements.txt
requests==2.30.0
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib